moved layout from root
