+++ src/app/(auth)/layout.tsx
@@ -4,6 +4,6 @@ interface LayoutProps {
   children: React.ReactNode
 }
 
-export default function RootLayout({ children }: LayoutProps): JSX.Element {
+export default function AuthLayout({ children }: LayoutProps): JSX.Element {
   return <React.Fragment>{children}</React.Fragment>
 }
